feat: add auth to post mutation

+++ src/modules/post/services/post.service.ts
@@ -1,16 +1,25 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { CreatePostInput } from '../dto/create-post.input';
 import { UpdatePostInput } from '../dto/update-post.input';
+import { RemovePostInput } from '../dto/remove-post.input';
+import { AssignTagInput } from '../dto/assign-tag.input';
+import { RemoveTagInput } from '../dto/remove-tag.input';
 import { PostRepository } from '../repositories/post.repository';
 import { Post } from '../entities/post.entity';
 
 @Injectable()
 export class PostService {
   constructor(private readonly postRepository: PostRepository) {}
-  async create(createPostInput: CreatePostInput): Promise<Post> {
+  async create(
+    createPostInput: CreatePostInput & { userId: string },
+  ): Promise<Post> {
     return await this.postRepository.create({
       ...createPostInput,
-      authorId: '7878c990-83c7-437e-a5db-2c133c6703d3',
+      authorId: createPostInput.userId,
     });
   }
 
@@ -28,19 +37,63 @@ export class PostService {
     return post;
   }
 
-  async update(id: string, updatePostInput: UpdatePostInput): Promise<Post> {
+  async update(
+    updatePostInput: UpdatePostInput & { userId: string },
+  ): Promise<Post> {
+    const { id, userId } = updatePostInput;
+    const post = await this.postRepository.findOne(id);
+
+    if (!post) {
+      throw new NotFoundException(`Post with id ${id} not found`);
+    }
+
+    if (post.authorId !== userId) {
+      throw new UnauthorizedException();
+    }
+
     return await this.postRepository.update(id, updatePostInput);
   }
 
-  async remove(id: string) {
+  async remove(removePostInput: RemovePostInput & { userId: string }) {
+    const { id, userId } = removePostInput;
+    const post = await this.postRepository.findOne(id);
+
+    if (!post) {
+      throw new NotFoundException(`Post with id ${id} not found`);
+    }
+
+    if (post.authorId !== userId) {
+      throw new UnauthorizedException();
+    }
     return await this.postRepository.delete(id);
   }
 
-  async assignTag(id: string, tagId: number) {
+  async assignTag(assignTagInput: AssignTagInput & { userId: string }) {
+    const { id, tagId, userId } = assignTagInput;
+    const post = await this.postRepository.findOne(id);
+
+    if (!post) {
+      throw new NotFoundException(`Post with id ${id} not found`);
+    }
+
+    if (post.authorId !== userId) {
+      throw new UnauthorizedException();
+    }
     return await this.postRepository.assignTag(id, tagId);
   }
 
-  async removeTag(id: string, tagId: number) {
+  async removeTag(removeTagInput: RemoveTagInput & { userId: string }) {
+    const { id, tagId, userId } = removeTagInput;
+
+    const post = await this.postRepository.findOne(id);
+
+    if (!post) {
+      throw new NotFoundException(`Post with id ${id} not found`);
+    }
+
+    if (post.authorId !== userId) {
+      throw new UnauthorizedException();
+    }
     return await this.postRepository.removeTag(id, tagId);
   }
 }
